Add julia lu factorization. Fix some promotion for linear systems and fix bidiagonal and triangular solvers for multiple right hand side. As consequence, change \ for bitarray to return Array{Float32} and introduce convert methods for sparse matrices without specifying integer type. Devectorize matrix norm.

+++ test/bitarray.jl
@@ -288,8 +288,8 @@ end
 b2 = randbool(n1, n1)
 
 @check_bit_operation (*) Matrix{Int} (b1, b2)
-@check_bit_operation (/) Matrix{Float64} (b1, b1)
-@check_bit_operation (\) Matrix{Float64} (b1, b1)
+@check_bit_operation (/) Matrix{Float32} (b1, b1)
+@check_bit_operation (\) Matrix{Float32} (b1, b1)
 
 b0 = falses(0)
 @check_bit_operation (&) BitVector (b0, b0)
